refactor: separate responsibilities to comply with SRP
- Moved WebSocket-related logic from ChatController to WebSocketController
- Handled chat room CRUD operations in ChatRoomController
- ChatController now focuses solely on real-time message handling
