+++ src/main/java/com/devcard/devcard/chat/controller/ChatController.java
@@ -1,13 +1,27 @@
 package com.devcard.devcard.chat.controller;
 
-import org.springframework.web.bind.annotation.GetMapping;
+import com.devcard.devcard.chat.dto.SendingMessageRequest;
+import com.devcard.devcard.chat.dto.SendingMessageResponse;
+import com.devcard.devcard.chat.service.ChatService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/chat")
 public class ChatController {
 
-    @GetMapping("/chat")
-    public String chatGET() {
-        return "chat";
+    private final ChatService chatService;
+
+    public ChatController(ChatService chatService) {
+        this.chatService = chatService;
+    }
+
+    // 1. 메세지 전송
+    @PostMapping("")
+    public ResponseEntity<SendingMessageResponse> sendMessage(@RequestBody SendingMessageRequest sendingMessageRequest) {
+        return ResponseEntity.status(201).body(chatService.sendMessage(sendingMessageRequest));
     }
 }
