+++ src/main/java/com/devcard/devcard/chat/controller/ChatRoomController.java
@@ -4,8 +4,6 @@ import com.devcard.devcard.chat.dto.ChatRoomListResponse;
 import com.devcard.devcard.chat.dto.ChatRoomResponse;
 import com.devcard.devcard.chat.dto.CreateRoomRequest;
 import com.devcard.devcard.chat.dto.CreateRoomResponse;
-import com.devcard.devcard.chat.dto.SendingMessageRequest;
-import com.devcard.devcard.chat.dto.SendingMessageResponse;
 import com.devcard.devcard.chat.service.ChatRoomService;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
@@ -33,11 +31,6 @@ public class ChatRoomController {
         return ResponseEntity.status(201).body(chatRoomService.createChatRoom(createRoomRequest));
     }
 
-    // 1. 메세지 전송
-    @PostMapping("")
-    public ResponseEntity<SendingMessageResponse> sendMessage(@RequestBody SendingMessageRequest sendingMessageRequest) {
-        return ResponseEntity.status(201).body(chatRoomService.sendMessage(sendingMessageRequest));
-    }
 
     // 2. 전체 채팅 목록 조회
     @GetMapping("")
