+++ src/main/java/com/devcard/devcard/chat/handler/ChatHandler.java
@@ -1,6 +1,7 @@
 package com.devcard.devcard.chat.handler;
 
 import com.devcard.devcard.chat.service.ChatRoomService;
+import com.devcard.devcard.chat.service.ChatService;
 import java.io.IOException;
 import java.util.List;
 import org.springframework.stereotype.Component;
@@ -12,20 +13,21 @@ import org.springframework.web.socket.handler.TextWebSocketHandler;
 @Component
 public class ChatHandler extends TextWebSocketHandler {
 
-    // service 계층 사용
     private final ChatRoomService chatRoomService;
+    private final ChatService chatService;
 
     // 의존성 주입
-    public ChatHandler(ChatRoomService chatRoomService) {
+    public ChatHandler(ChatRoomService chatRoomService, ChatService chatService) {
         this.chatRoomService = chatRoomService;
+        this.chatService = chatService;
     }
 
     @Override
     protected void handleTextMessage(WebSocketSession session, TextMessage textMessage) throws Exception {
         String payload = textMessage.getPayload();
-        String chatId = chatRoomService.extractChatId(payload);
-        String message = chatRoomService.extractMessage(payload);
-        List<WebSocketSession> chatRoom = chatRoomService.getChatRoomSessions(chatId);
+        String chatId = chatService.extractChatId(payload);
+        String message = chatService.extractMessage(payload);
+        List<WebSocketSession> chatRoom = chatService.getChatRoomSessions(chatId);
 
         if (chatRoom != null) { // 채팅방이 있을 때
             for (WebSocketSession webSocketSession : chatRoom) { // 해당 채팅방의 모든 세션에 대해서 메세지 보내기
@@ -43,20 +45,20 @@ public class ChatHandler extends TextWebSocketHandler {
     // 클라이언트 접속 시 호출=> id를 추출하여 채팅방이 이미 존재하면 기존에 추가 없다면 생성후 추가
     @Override
     public void afterConnectionEstablished(WebSocketSession session) {
-        String chatId = chatRoomService.extractChatIdFromSession(session);
+        String chatId = chatService.extractChatIdFromSession(session);
         // 채팅방 존재 여부 확인
         if (!chatRoomService.existsChatRoom(chatId)) {// 채팅방이 존재 하지 않을 시 DB에 채팅방 생성
             // chatRoomService.createChatRoom(new CreateRoomRequest(chatId)); => 어떻게 참여자 id를 가져와 추가할 것인지 생각후
         }
         // 세션 추가
-        chatRoomService.addSessionToChatRoom(chatId, session);
+        chatService.addSessionToChatRoom(chatId, session);
     }
 
     // 클라이언트 접속 해제 시 호출
     @Override
     public void afterConnectionClosed(WebSocketSession session, CloseStatus status) {
-        String chatId = chatRoomService.extractChatIdFromSession(session);
-        List<WebSocketSession> sessions = chatRoomService.getChatRoomSessions(chatId);
+        String chatId = chatService.extractChatIdFromSession(session);
+        List<WebSocketSession> sessions = chatService.getChatRoomSessions(chatId);
         if (sessions != null) {
             sessions.remove(session);
             // 채팅방을 제거하지 않고 해당 세션만 제거
