+++ src/main/java/com/devcard/devcard/chat/repository/ChatRepository.java
@@ -5,7 +5,7 @@ import com.devcard.devcard.chat.model.ChatRoom;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+public interface ChatRepository extends JpaRepository<ChatMessage, Long> {
 
     List<ChatMessage> findByChatRoomOrderByTimestampAsc(ChatRoom chatRoom);
